% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacons.R
\docType{data}
\name{beacons}
\alias{beacons}
\title{Middle Earth beacon data}
\format{
\subsection{\code{beacons}}{

A data frame with 150 rows and 2 columns:
\describe{
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
beacons
}
\description{
Geospatial shapefiles of beacons throughout Middle Earth
Report ...
}
\keyword{datasets}
