% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{citadels}
\alias{citadels}
\title{Middle Earth citadels data}
\format{
\subsection{\code{citadels}}{

A data frame with 119 rows and 3 columns:
\describe{
\item{NAME}{citadel name}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
citadels
}
\description{
Geospatial shapefiles of citadels throughout Middle Earth
Report ...
}
\keyword{datasets}
