% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\docType{data}
\name{cities}
\alias{cities}
\title{Middle Earth cities data}
\format{
\subsection{\code{cities}}{

A data frame with 30 rows and 4 columns:
\describe{
\item{TYPE}{city type}
\item{NAME}{city name}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
cities
}
\description{
Geospatial shapefiles of cities throughout Middle Earth
Report ...
}
\keyword{datasets}
