% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{coastline}
\alias{coastline}
\title{Middle Earth coastline data}
\format{
\subsection{\code{coastline}}{

A data frame with 110 rows and 14 columns:
\describe{
\item{LENGTH}{length of coastline}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
coastline
}
\description{
Geospatial shapefiles of coastlines throughout Middle Earth
Report ...
}
\keyword{datasets}
