% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_placenames.R
\docType{data}
\name{combined_placenames}
\alias{combined_placenames}
\title{Middle Earth regional name data}
\format{
\subsection{\code{combined_placenames}}{

A data frame with 785 rows and 11 columns:
\describe{
\item{LAYER}{region type}
\item{NAME}{name of region}
\item{FONT_NAME}{name of font used}
\item{FONT_COLOR}{color of font used}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
combined_placenames
}
\description{
Geospatial shapefiles of regional names throughout Middle Earth
Report ...
}
\keyword{datasets}
