% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\docType{data}
\name{contours}
\alias{contours}
\title{Middle Earth contour lines data}
\format{
\subsection{\code{contours}}{

A data frame with 4266 rows and 14 columns:
\describe{
\item{LENGTH}{contour length}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
contours
}
\description{
Geospatial shapefiles of countour lines throughout Middle Earth
Report ...
}
\keyword{datasets}
