% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forests.R
\docType{data}
\name{forests}
\alias{forests}
\title{Middle Earth forest data}
\format{
\subsection{\code{forests}}{

A data frame with 2033 rows and 8 columns:
\describe{
\item{NAME}{forest name}
\item{TYPE}{forest type}
\item{PERIMETER}{forest perimeter}
\item{ENCLOSED_A}{forest area}
\item{ISLAND_ARE}{island area}
\item{NUMBER}{forest data ID number}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
forests
}
\description{
Geospatial shapefiles of forests throughout Middle Earth
Report ...
}
\keyword{datasets}
