% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{hamlets}
\alias{hamlets}
\title{Middle Earth hamlet data}
\format{
\subsection{\code{hamlets}}{

A data frame with 20 rows and 4 columns:
\describe{
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
hamlets
}
\description{
Geospatial shapefiles of hamlets (small settlements) throughout Middle Earth
Report ...
}
\keyword{datasets}
