% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_contours.R
\docType{data}
\name{helper_contours}
\alias{helper_contours}
\title{Middle Earth helper contour line data}
\format{
\subsection{\code{helper_contours}}{

A data frame with 19 rows and 8 columns:
\describe{
\item{LENGTH}{contour line length}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
helper_contours
}
\description{
Geospatial shapefiles of helper contour lines throughout Middle Earth
Report ...
}
\keyword{datasets}
