% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hills.R
\docType{data}
\name{hills}
\alias{hills}
\title{Middle Earth hill data}
\format{
\subsection{\code{hills}}{

A data frame with 2179 rows and 7 columns:
\describe{
\item{PERIMETER}{hill perimeter}
\item{ENCLOSED_A}{enclosed hill area}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
hills
}
\description{
Geospatial shapefiles of hills throughout Middle Earth
Report ...
}
\keyword{datasets}
