% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{lakes}
\alias{lakes}
\title{Middle Earth lakes data}
\format{
\subsection{\code{lakes}}{

A data frame with 256 rows and 7 columns:
\describe{
\item{NAME}{lake name}
\item{PERIMETER}{lake perimeter}
\item{ENCLOSED_A}{lake area}
\item{ISLAND_ARE}{island area}
\item{type}{type}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
lakes
}
\description{
Geospatial shapefiles of lakes throughout Middle Earth
Report ...
}
\keyword{datasets}
