% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meters_to_miles.R
\name{meters_to_miles}
\alias{meters_to_miles}
\title{Convert meters to miles}
\usage{
meters_to_miles(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
a numeric value
}
\description{
Convert meters to miles
}
\examples{
meters_to_miles(500)
}
