% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miles_to_meters.R
\name{miles_to_meters}
\alias{miles_to_meters}
\title{Convert miles to meters}
\usage{
miles_to_meters(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
a numeric value
}
\description{
Convert miles to meters
}
\examples{
miles_to_meters(500)
}
