% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{mountains}
\alias{mountains}
\title{Middle Earth mountains data}
\format{
\subsection{\code{mountains}}{

A data frame with 10 rows and 3 columns:
\describe{
\item{name}{mountain range name}
\item{fontsizeKm}{fontzize in km}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
mountains
}
\description{
Geospatial shapefiles of mountains throughout Middle Earth
Report ...
}
\keyword{datasets}
