% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{placenames}
\alias{placenames}
\title{Middle Earth place name data}
\format{
\subsection{\code{placenames}}{

A data frame with 785 rows and 11 columns:
\describe{
\item{LAYER}{place name type}
\item{NAME}{place name}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
placenames
}
\description{
Geospatial shapefiles of place names throughout Middle Earth
Report ...
}
\keyword{datasets}
