% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Middle Earth regions data}
\format{
\subsection{\code{regions}}{

A data frame with 14 rows and 3 columns:
\describe{
\item{name}{region name}
\item{fontsizeKm}{font size in km}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
regions
}
\description{
Geospatial shapefiles of regions throughout Middle Earth
Report ...
}
\keyword{datasets}
