% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{Middle Earth rivers data}
\format{
\subsection{\code{rivers}}{

A data frame with 2497 rows and 7 columns:
\describe{
\item{LENGTH}{river length}
\item{TYPE}{river type}
\item{NAME}{river name}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
rivers
}
\description{
Geospatial shapefiles of rivers throughout Middle Earth
Report ...
}
\keyword{datasets}
