% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roads.R
\docType{data}
\name{roads}
\alias{roads}
\title{Middle Earth road data}
\format{
\subsection{\code{roads}}{

A data frame with 1010 rows and 4 columns:
\describe{
\item{NAME}{road name}
\item{TYPE}{road type}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
roads
}
\description{
Geospatial shapefiles of roads throughout Middle Earth
Report ...
}
\keyword{datasets}
