% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruins.R
\docType{data}
\name{ruins}
\alias{ruins}
\title{Middle Earth ruins data}
\format{
\subsection{\code{ruins}}{

A data frame with 237 rows and 3 columns:
\describe{
\item{NAME}{ruin name}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
ruins
}
\description{
Geospatial shapefiles of ruins throughout Middle Earth
Report ...
}
\keyword{datasets}
