% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{stonefields}
\alias{stonefields}
\title{Middle Earth stonefields data}
\format{
\subsection{\code{stonefields}}{

A data frame with 7 rows and 6 columns:
\describe{
\item{PERIMETER}{stonefield perimeter}
\item{ENCLOSED_A}{stonefield enclosed area}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
stonefields
}
\description{
Geospatial shapefiles of stonefields throughout Middle Earth
Report ...
}
\keyword{datasets}
