% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towers_and_keeps.R
\docType{data}
\name{towers_and_keeps}
\alias{towers_and_keeps}
\title{Middle Earth towers and keeps data}
\format{
\subsection{\code{towers_and_keeps}}{

A data frame with 202 rows and 3 columns:
\describe{
\item{NAME}{tower or keep name}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
towers_and_keeps
}
\description{
Geospatial shapefiles of towers and keeps throughout Middle Earth
Report ...
}
\keyword{datasets}
