% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{towns}
\alias{towns}
\title{Middle Earth towns data}
\format{
\subsection{\code{towns}}{

A data frame with 425 rows and 4 columns:
\describe{
\item{TYPE}{town type}
\item{NAME}{toen name}
\item{ORIGIN}{data source}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
towns
}
\description{
Geospatial shapefiles of towns throughout Middle Earth
Report ...
}
\keyword{datasets}
