% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{vulcanism}
\alias{vulcanism}
\title{Middle Earth vulcanism data}
\format{
\subsection{\code{vulcanism}}{

A data frame with 110 rows and 7 columns:
\describe{
\item{NAME}{volcano name}
\item{PERIMETER}{volcano perimeter}
\item{ENCLOSED_A}{volcano area}
\item{ISLAND_ARE}{island area}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
vulcanism
}
\description{
Geospatial shapefiles of volcanos throughout Middle Earth
Report ...
}
\keyword{datasets}
