% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wetlands.R
\docType{data}
\name{wetlands}
\alias{wetlands}
\title{Middle Earth wetlands data}
\format{
\subsection{\code{wetlands}}{

A data frame with 181 rows and 7 columns:
\describe{
\item{NAME}{wetlands name}
\item{PERIMETER}{wetlands perimeter}
\item{ENCLOSED_A}{wetlands area}
\item{geometry}{geometry}
...
}
}
}
\source{
\url{https://github.com/jvangeld/ME-GIS}
}
\usage{
wetlands
}
\description{
Geospatial shapefiles of wetlands throughout Middle Earth
Report ...
}
\keyword{datasets}
